% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{angle_in_range}
\alias{angle_in_range}
\title{Check whether an angle is within a specified range}
\usage{
angle_in_range(x, lower, upper)
}
\arguments{
\item{x}{Numeric value indicating the angle to be checked, in units of
degrees on a 360 degree scale. May be of class \code{circular} or
\code{numeric}.}

\item{lower}{Numeric value indicating the lower end of the angular range to
be evaluated. Units, range, and class as with \code{x}.}

\item{upper}{Numeric value indicating the upper end of the angular range to
be evaluated. Units, range, and class as with \code{x}.}
}
\value{
Logical value indicating whether \code{x} is within the interval
  \code{c(lower, upper)}.
}
\description{
Helper function used by \code{\link{BaBA_caribou}} to check whether a given
angle is within the specified range of values. Code adapted from
\href{https://stackoverflow.com/questions/66799475/how-to-elegantly-find-if-an-angle-is-between-a-range}{StackOverflow}.
}
