% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_angle}
\alias{calc_angle}
\title{Calculate headings of one or more linear steps}
\usage{
calc_angle(x)
}
\arguments{
\item{x}{\code{\href{https://cran.r-project.org/package=sf}{sf}}
\code{POINT} object representing points of movement locations
during an encounter or segment points along a linear barrier. This should
have columns labelled \code{x} and \code{y} indicating the xy coordinates
of feature \code{x}.}
}
\value{
Numeric vector with length equal to \code{nrow(x)} indicating the
  absolute angles in radians between each pair of points and the x-axis. The
  final value will always be NA as an angle cannot be calculated for the last
  point in a sequence.
}
\description{
Helper function used by \code{\link{BaBA_caribou}} to calculate absolute
angles between a movement step or barrier segment and the x-axis to provide
an indication of heading. This is adapted from code in the \code{as.ltraj}
function of
\code{\href{https://cran.r-project.org/package=adehabitatLT}{adehabitatLT}}.
}
