% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_dist}
\alias{calc_dist}
\title{Helper function for calculating Euclidean distance between locations}
\usage{
calc_dist(x.start, x.end, y.start, y.end)
}
\arguments{
\item{x.start}{x-coordinate of the starting location.}

\item{x.end}{x-coordinate of the ending location.}

\item{y.start}{y-coordinate of the starting location.}

\item{y.end}{y-coordinate of the ending location.}
}
\value{
Numeric value indicating the Euclidean distance (in units of the
  telemetry locations, often meters or degrees) between the input locations.
}
\description{
\code{calc_dist} is a helper function used by the \code{\link{BaBA}} function
to simplify calculation of Euclidean distance between subsequent telemetry
locations.
}
\details{
The coordinate reference system for the arguments is assumed to be
  the same as that for all location data in the \code{animal} argument of
  \code{\link{BaBA_caribou}}, matching the \code{crs} argument. Euclidean
  distance is simply calculated using the Pythagorean theorem and does not
  take into account the curvature of the earth. It is thus most appropriate
  for use at local scales and may not be valid if the time difference between
  locations is excessively large and the animal exhibits large displacement
  between locations.
}
