% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{line_extend}
\alias{line_extend}
\title{Project a point along an angle and distance}
\usage{
line_extend(pt, angle, len)
}
\arguments{
\item{pt}{\code{data.frame} with one row and columns \code{x} and \code{y}
indicating the location of the point to be projected into space.}

\item{angle}{Numerical value indicating the angle, in units of degrees,
indicating the direction along which \code{pt} should be projected.}

\item{len}{Numerical value indicating the distance which {pt} should be
moved, in the same units as \code{pt}.}
}
\value{
\code{data.frame} with two rows and two columns. Columns indicate the
  x and y coordinates of the points. The first row represents the location of
  \code{pt}. The second row indicates the location of the newly projected
  point \code{len} units away from \code{pt} at a bearing of \code{angle}.
}
\description{
Helper function used by \code{\link{BaBA_caribou}} to project a point
elsewhere in space given an initial point, angle for projection, and distance
to move the point. This is adapted from the \code{CreateSegmentAngle}
function in the
\href{https://cran.r-project.org/package=LearnGeom}{\code{LearnGeom}}
package.
}
