% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{strtns}
\alias{strtns}
\title{Calculate straightness of movement segment}
\usage{
strtns(mov_seg)
}
\arguments{
\item{mov_seg}{\code{\href{https://cran.r-project.org/package=sf}{sf}}
\code{POINT} object representing locations that are part of the encounter
being analyzed.}
}
\value{
Numeric value indicating the straightness of movement during the
  encounter. Values range between 0-1, with values closer to 0 indicating
  more sinuous movement.
}
\description{
Helper function used by the \code{\link{BaBA_caribou}} function to calculate
straightness of movement during an encounter as the ratio of Euclidean to
path distance.
}
